requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.9.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
coloredlogs==15.0.1
cryptography==41.0.4
dataclasses-json==0.6.0
distlib==0.3.7
emoji==2.8.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
identify==2.5.29
idna==3.4
importlib-resources==6.1.0
isort==5.12.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.301
langsmith==0.0.40
lxml==4.9.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.6
numpy==1.24.4
onnxruntime==1.16.0
openai==0.28.1
overrides==7.4.0
packaging==23.1
pathspec==0.11.2
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
Pillow==10.0.1
platformdirs==3.10.0
posthog==3.0.2
pre-commit==3.4.0
protobuf==4.24.3
pulsar-client==3.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.0
pydantic_core==2.10.0
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.66.1
transformers==4.33.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.16
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.5
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
